refactor: 패키지 이동 및 리팩토링
