+++ tomcat/src/main/java/nextstep/jwp/controller/HomeHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.controller;
 
 import static org.apache.coyote.http11.http.HttpVersion.HTTP11;
 import static org.apache.coyote.http11.header.ContentType.HTML;
