+++ tomcat/src/main/java/nextstep/jwp/controller/resource/ResourceUrls.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler.resource;
+package nextstep.jwp.controller.resource;
 
 abstract public class ResourceUrls {
